backend/requirements.txt
django==4.2.5
djangorestframework==3.14.0
django-filter==23.3
djangorestframework-simplejwt==5.3.0
Pillow==9.5.0
python-dotenv==1.0.0
Faker==19.6.1